Add pts-init alias. Make nose aliases conditional

+++ .zshrc
@@ -53,10 +53,14 @@ fi
 alias ll="ls -alG"
 alias brew-up="brew update && brew upgrade -g && brew autoremove && brew cleanup -s --prune=all && brew outdated --cask"
 
-alias node="docker run --rm -it -v .:/src -w /src node $@"
-alias npm="node npm $@"
-alias npx="node npx $@"
-alias npr="npm run $@"
+# node, typescript
+if ! type pnpm &> /dev/null; then
+    alias node="docker run --rm -it -v .:/src -w /src node $@"
+    alias npm="node npm $@"
+    alias npx="node npx $@"
+    alias npr="npm run $@"
+fi
+
 alias tsr="node -r esbuild-register $@"
 
 alias tsconfig-init='cat > tsconfig.json <<EOF
@@ -74,3 +78,7 @@ EOF
 '
 
 alias ts-init="npm init -y && npm add -D typescript @types/node esbuild esbuild-register @tsconfig/strictest && npx tsc --init && mkdir src && echo 'console.log(\"Hello, world!\");' > src/index.ts"
+
+if type pnpm &> /dev/null; then
+    alias pts-init="pnpm init && pnpm add -D typescript esbuild esbuild-register @types/node @tsconfig/strictest && pnpm tsc --init && mkdir src && echo 'console.log(\"Hello, world!\");' >> src/index.ts"
+fi
